% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposome_example}
\alias{exposome_example}
\title{Example exposome dataset for ExpoMet}
\format{
A data frame with rows as participants and columns including:
\describe{
\item{exp_id}{The ID for the participants.}
\item{C0001}{The values for exposomic feature 'C0001'.}
}
}
\usage{
exposome_example
}
\description{
A de-identified example dataset used to demonstrate
the main ExpoMet workflow.
}
\keyword{datasets}
