% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_feature_key_final}
\alias{simulated_feature_key_final}
\title{Example dataset of the feature info for ExpoMet, associated with the "exposome_example" file}
\format{
A data frame with rows as exposomic features, and columns are the associated info:
\describe{
\item{feature_name}{Correspond to the column names in "exposome_example"}
}
}
\usage{
simulated_feature_key_final
}
\description{
Example dataset of the feature info for ExpoMet, associated with the "exposome_example" file
}
\keyword{datasets}
