% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno_example}
\alias{pheno_example}
\title{Example phenotype dataset for ExpoNet}
\format{
A data frame with rows as participants and columns including:
\describe{
\item{exp_id}{The ID for the participants.}
\item{sex}{Participant sex, coded as 1 or 2.}
\item{age}{Participant age in years, ranging from 18 to 80.}
\item{smoking}{Smoking status, coded as 1, 2, or 3.}
\item{Disease}{Disease status, coded as "Yes" or "No".}
}
}
\usage{
pheno_example
}
\description{
A simulated phenotype dataset used to demonstrate
the main ExpoNet workflow.
}
\keyword{datasets}
