% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview of Data}
\usage{
preview(omic_features, id_col = "exp_id")
}
\arguments{
\item{omic_features}{A dataframe with an id column and exposure columns.}

\item{id_col}{A string specifying the name of the id column.}
}
\value{
A printed summary containing:

\describe{
\item{Duplicate IDs}{identifies any duplicate IDs present in the ID column.}
\item{Non-numeric columns}{the exposure columns that contain non-numeric values
and the percentage of columns in the dataset with non-numeric values.}
\item{Missing data}{the exposure columns and rows that contain missing values (\code{NA}
or \code{NaN}) and the percentage of missing values per column or row.}
\item{Zeroes Present}{the exposure columns containing zeroes and the percentage
of missing values per column.}
\item{Zero variance}{the exposure columns with zero variance across all samples.}
\item{Skewness}{the exposure columns with skewness that significantly deviates from
a normal distribution, evaluated using the D'Agostino-Pearson test.}
\item{Minimum Inflation}{the exposure columns that have an overrepresented minimum value
of more than five times what is expected by chance.}
}
}
\description{
\code{preview} can be used as a quick overview of common data issues that may
affect model fitting and is designed to be used prior to any subsequent
function in the ExpoMet package. This function checks and identifies
duplicate IDs, non-numeric columns, data missingness, negative numbers, exposures with zeroes,
exposures with zero variance, exposures with significantly skewed data, using
the  D'Agostino-Pearson test (FDR < 0.05), and whether the minimum value in the dataset is
overrepresented by more than five times what would be expected by chance.
The first 5 features are returned for each check, with the option to view the
entire dataset.
}
\examples{
\dontrun{
preview(omic_features, id_col = "exp_id")
}

}
