% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_feature_key_final}
\alias{simulated_feature_key_final}
\title{Simulated feature annotation key}
\format{
A data frame containing simulated exposomic feature identifiers
and associated feature annotation information.
}
\usage{
simulated_feature_key_final
}
\description{
A simulated annotation table corresponding to features in the
\code{exposome_example} dataset.
}
\keyword{datasets}
